+++ LearningJava/src/Java_Basics/basics/Array_of_Objects.java
@@ -1,4 +1,4 @@
-package Java_Basics;
+package Java_Basics.basics;
 
 class Students{
     int marks;
