+++ LearningJava/src/Java_Basics/basics/Main.java
@@ -1,4 +1,4 @@
-package Java_Basics;
+package Java_Basics.basics;
 public class Main {
     public static void main(String[] args) {
         int n = 1;
